+++ src/extension/actions.js
@@ -25,7 +25,7 @@ import {
   detectLegacySidekick,
   updateProject as updateProjectConfig,
 } from './project.js';
-import { ADMIN_ORIGIN, createAdminUrl } from './utils/admin.js';
+import { ADMIN_ORIGIN, ADMIN_ORIGIN_NEW, createAdminUrl } from './utils/admin.js';
 import { getConfig } from './config.js';
 import { getDisplay, setDisplay } from './display.js';
 import { urlCache } from './url-cache.js';
@@ -55,7 +55,8 @@ async function updateAuthToken({
 }, { tab }) {
   if (owner) {
     try {
-      if (new URL(tab.url).origin === ADMIN_ORIGIN
+      const { origin } = new URL(tab.url);
+      if ((origin === ADMIN_ORIGIN || origin === ADMIN_ORIGIN_NEW)
         && authToken !== undefined) {
         await setAuthToken(
           owner,
@@ -115,6 +116,7 @@ export async function showSidekickNotification(tabId, data, callback) {
 function isTrustedOrigin(origin) {
   const TRUSTED_ORIGINS = [
     ADMIN_ORIGIN,
+    ADMIN_ORIGIN_NEW,
     'https://labs.aem.live',
     'https://tools.aem.live',
     'http://localhost:3000',
@@ -416,7 +418,9 @@ async function importProjects(tab) {
     {
       message: chrome.i18n.getMessage(i18nKey, `${imported}`),
       headline: chrome.i18n.getMessage('config_project_import_headline'),
-    });
+    },
+    notificationConfirmCallback(tab.id),
+  );
 }
 
 /**
@@ -425,7 +429,7 @@ async function importProjects(tab) {
  */
 async function manageProjects(tab) {
   await chrome.tabs.create({
-    url: 'https://labs.aem.live/tools/project-admin/index.html',
+    url: 'https://tools.aem.live/tools/project-admin/index.html',
     openerTabId: tab.id,
     windowId: tab.windowId,
   });