+++ src/extension/app/store/bulk.js
@@ -159,6 +159,7 @@ export class BulkStore {
     return [...document.querySelectorAll('#appRoot [aria-selected="true"]:not([aria-checked="true"]')]
       // exclude folders
       .filter((row) => !row.querySelector('img')?.getAttribute('src').includes('/foldericons/')
+        && !row.getAttribute('type')?.includes('button')
         && !row.querySelector('img')?.getAttribute('src').endsWith('folder.svg')
         && !row.querySelector('svg')?.parentElement.className.toLowerCase().includes('folder'))
       // extract file name and type
@@ -200,7 +201,10 @@ export class BulkStore {
    * @returns {BulkSelection} The selection
    */
   #getGoogleDriveBulkSelection(document) {
-    return [...document.querySelectorAll('#drive_main_page [aria-selected="true"]')]
+    return [
+      ...document.querySelectorAll('div[role="main"] [aria-selected="true"]'),
+      ...document.querySelectorAll('#drive_main_page [aria-selected="true"]'), // legacy
+    ]
       // extract file name and type
       .map((row) => {
         const file = (row.querySelector(':scope td div[data-id] > span > strong') // list layout