+++ test/app/store/app.test.js
@@ -104,7 +104,6 @@ describe('Test App Store', () => {
     await testDefaultConfig();
 
     expect(appStore.siteStore.plugins.length).to.eq(9);
-    expect(appStore.siteStore.scriptUrl).to.eq('https://www.hlx.live/tools/sidekick/index.js');
     expect(appStore.siteStore.host).to.eq('www.aemboilerplate.com');
     expect(appStore.siteStore.innerHost).to.eq('custom-preview-host.com');
     expect(appStore.siteStore.liveHost).to.eq('custom-live-host.com');